fix compile errors

+++ FSPprj/user_src/user.h
@@ -8,7 +8,7 @@
 #include "motor.h"
 #include "interface_driver.h"
 #include "screen_uart.h"
-
+#include "fire_alarm.h"
 
 #define RunForward 0
 #define RunBackward 1
@@ -37,7 +37,7 @@ extern float speed_forward,target_angle;
 extern volatile uint8_t onworking,run_mode;
 void uprintf(uart_ctrl_t *p_ctrl,char *fmt, ...);
 void go_to_line1(float speed);
-void set_position(uint32_t pos);
+void set_position(uint16_t pos);
 
 
 #endif
